+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Invalid keyword in documentation stating that transaction ends with `STOP` but
+  it is signaled by keyword `END` instead
+
+
 ## [70.0.0] - 22-01-26
 ### Added
 - Change log that leaves minimal trace of changes updater performed stored in