+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Provide function `get_turris_serial` returning serial number available inside Lua
+
 ## [70.0.2] - 22-11-02
 ### Changed
 - Use escape sequences only on terminal